Render views of the overall galaxy with stars
Still need to reimplement rendering the jump lanes and sector boundaries

+++ SpaceOpera/Core/Universe/Galaxy.cs
@@ -2,10 +2,10 @@ namespace SpaceOpera.Core.Universe
 {
     public class Galaxy
     {
-        public double Radius { get; }
+        public float Radius { get; }
         public List<StarSystem> Systems { get; }
 
-        public Galaxy(double radius, IEnumerable<StarSystem> systems)
+        public Galaxy(float radius, IEnumerable<StarSystem> systems)
         {
             Radius = radius;
             Systems = systems.ToList();
